wip: rearrange location
